requirements-dev.txt
black==22.6.0
virtualenv==20.13.0
djangorestframework-stubs==1.7.0
django-stubs==1.12.0
Faker==15.3.2
